requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
chromadb==0.4.22
pypdf==4.0.0
streamlit==1.30.0
python-dotenv==1.0.0
openai==1.10.0
tiktoken==0.5.2